+++ src/collections/btree/btreemap.rs
@@ -251,16 +251,18 @@ where
         }
     }
 
-    #[cfg(test)]
-    fn correct_empty_state(&self) -> bool {
+    // Function used in tests to enforce invariants below
+
+    #[doc(hidden)]
+    pub fn correct_empty_state(&self) -> bool {
         match &self.0 {
             BTreeMapInner::LeafNode(leaves) => leaves.len() == 0,
             _ => false,
         }
     }
 
-    #[cfg(test)]
-    fn all_leaves_at_same_level(&self) -> bool {
+    #[doc(hidden)]
+    pub fn all_leaves_at_same_level(&self) -> bool {
         use crate::All;
 
         match self.walk(All) {
@@ -280,120 +282,6 @@ where
                 }
             }
             None => true,
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::{MaxKey, TreeViz};
-
-    use super::BTreeMap;
-
-    use rkyv::rend::LittleEndian;
-
-    const S: i32 = 4;
-    const N: i32 = 16;
-
-    #[test]
-    fn btree_add_remove_simple() {
-        let mut map =
-            BTreeMap::<LittleEndian<i32>, i32, MaxKey<LittleEndian<i32>>>::new(
-            );
-
-        for o in S..N {
-            println!("\n------------\nTESTING N = {}", o);
-
-            for i in 0..o {
-                println!("insert {:?}", i);
-                assert_eq!(map.insert(LittleEndian::from(i), i), None);
-
-                map.print_tree();
-
-                assert!(map.all_leaves_at_same_level());
-            }
-
-            for i in 0..o {
-                println!("removing {:?}", i);
-
-                assert_eq!(map.remove(&LittleEndian::from(i)), Some(i));
-
-                map.print_tree();
-
-                assert!(map.all_leaves_at_same_level());
-            }
-
-            assert!(map.correct_empty_state());
-        }
-    }
-
-    #[test]
-    fn btree_add_remove_reverse() {
-        let mut map =
-            BTreeMap::<LittleEndian<i32>, i32, MaxKey<LittleEndian<i32>>>::new(
-            );
-
-        for o in S..N {
-            for i in 0..o {
-                let i = o - i - 1;
-                assert_eq!(map.insert(LittleEndian::from(i), i), None);
-
-                assert!(map.all_leaves_at_same_level());
-            }
-
-            for i in 0..o {
-                let i = o - i - 1;
-                assert_eq!(map.remove(&LittleEndian::from(i)), Some(i));
-                println!("removed {}", i);
-                map.print_tree();
-
-                assert!(map.all_leaves_at_same_level());
-            }
-        }
-
-        assert!(map.correct_empty_state());
-    }
-
-    #[test]
-    fn btree_add_change_remove() {
-        let mut map =
-            BTreeMap::<LittleEndian<i32>, i32, MaxKey<LittleEndian<i32>>>::new(
-            );
-
-        for o in S..N {
-            println!("\n------------\nTESTING N = {}", o);
-
-            for i in 0..o {
-                println!("insert {:?}", i);
-                assert_eq!(map.insert(LittleEndian::from(i), i), None);
-
-                map.print_tree();
-
-                assert!(map.all_leaves_at_same_level());
-            }
-
-            for i in 0..o {
-                println!("re-insert {:?}", i);
-                assert_eq!(map.insert(LittleEndian::from(i), i + 1), Some(i));
-
-                map.print_tree();
-            }
-
-            for i in 0..o {
-                assert_eq!(map.get(&LittleEndian::from(i)), Some(&(i + 1)));
-            }
-
-            for i in 0..o {
-                println!("removing {:?}", i);
-
-                assert_eq!(map.remove(&LittleEndian::from(i)), Some(i + 1));
-
-                map.print_tree();
-
-                assert!(map.all_leaves_at_same_level());
-            }
-
-            assert!(map.correct_empty_state());
         }
     }
 }
