+++ src/collections/mod.rs
@@ -1,2 +1,4 @@
 /// Microkelvin BTree implementation
 pub mod btree;
+
+pub use btree::BTreeMap;
